Add sorting strategy via upload
Implemented a new strategy example based on sorting algorithms. It will update with others algorithms to show the force of the strategy pattern. It could be also use to see the velocity difference by using an algorithm instead of another one, or use a type of list instead of another one.